refactor(localdebug): rename provisionLocalResource to procisionLocal
